fix: Error CS0266 Cannot implicitly convert type 'int?' to 'int'.
-> add: cast

+++ sshhub/Action.cs
@@ -252,7 +252,7 @@ namespace sshhub
                     continue;
                 }
 
-                target.id = newId;
+                target.id = (int)newId;
                 break;
             }
